add sync process for agendaupdates clean procedure up import
